requirements.txt
Django==3.2.5
djangorestframework==3.12.4
django-cors-headers==3.7.0
channels==3.0.4
channels-redis==3.3.0
coloredlogs==15.0.1